+++ mrmustard/physics/fock.py
@@ -247,7 +247,14 @@ def fidelity(state_a, state_b, a_ket: bool, b_ket: bool) -> Scalar:
             math.sum(math.conj(b) * math.matvec(math.reshape(state_a, (len(b), len(b))), b))
         )
 
-    raise NotImplementedError("Fidelity between mixed states is not implemented yet.")
+    # mixed state
+    # Richard Jozsa (1994) Fidelity for Mixed Quantum States, Journal of Modern Optics, 41:12, 2315-2323, DOI: 10.1080/09500349414552171
+    return (
+        math.trace(
+            math.sqrtm(math.matmul(math.matmul(math.sqrtm(state_a), state_b), math.sqrtm(state_a)))
+        )
+        ** 2
+    )
 
 
 def number_means(tensor, is_dm: bool):